Move cryspy_parser.py to easydiffraction/calculators/cryspy/parser.py
